api/conector/requirements.txt
fastapi
uvicorn[standard]
requests
pyyaml
psycopg2-binary
python-dotenv

